+++ app/src/js/components/DatePickerHeader/DatePickerHeader.js
@@ -1,27 +1,36 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import Datepicker from '../Datepicker/Datepicker';
+import Datepicker from '../Datepicker/DatepickerRange';
 
-const DatePickerHeader = ({ heading, onChange }) => (
-  <div className='content__header'>
-    <div className='row'>
-      <ul className='datetimeheader'>
-        <li>
-          <div className='datetimeheader__content'>
-            <h1 className='heading--xlarge'>{heading}</h1>
-          </div>
-        </li>
-        <li>
-          <Datepicker hideWrapper={true} onChange={onChange} />
-        </li>
-      </ul>
+const DatePickerHeader = ({ heading, onChange, showDatePicker = true }) => (
+  <>
+    {showDatePicker ? <div className='content__header'>
+      <div className='row'>
+        <ul className='datetimeheader'>
+          <li>
+            <div className='datetimeheader__content'>
+              <h1 className='heading--xlarge'>{heading}</h1>
+            </div>
+          </li>
+          <li>
+            <Datepicker hideWrapper={true} onChange={onChange} />
+          </li>
+        </ul>
+      </div>
     </div>
-  </div>
+      : <div className='content__header'>
+        <div className='row'>
+          <h1 className='heading--xlarge'>{heading}</h1>
+        </div>
+      </div>
+    }
+  </>
 );
 
 DatePickerHeader.propTypes = {
+  heading: PropTypes.string,
   onChange: PropTypes.func,
-  heading: PropTypes.string
+  showDatePicker: PropTypes.bool
 };
 
 export default DatePickerHeader;