+++ app/src/js/components/ReconciliationReports/index.js
@@ -7,6 +7,7 @@ import withQueryParams from 'react-router-query-params';
 import Sidebar from '../Sidebar/sidebar';
 import { strings } from '../locale';
 import { getCount, listReconciliationReports } from '../../actions';
+import CreateReconciliationReport from './create';
 import ReconciliationReportList from './list';
 import ReconciliationReport from './reconciliation-report';
 import DatePickerHeader from '../DatePickerHeader/DatePickerHeader';
@@ -19,6 +20,8 @@ const ReconciliationReports = ({
   params,
   queryParams,
 }) => {
+  const { pathname } = location;
+  const showSidebar = pathname !== '/reconciliation-reports/create';
   const filteredQueryParams = filterQueryParams(queryParams);
 
   function query() {
@@ -35,21 +38,28 @@ const ReconciliationReports = ({
       <Helmet>
         <title> Reconcilation Reports </title>
       </Helmet>
-      <DatePickerHeader onChange={query} heading={strings.reconciliation_reports} />
+      <DatePickerHeader onChange={query} heading={strings.reconciliation_reports} showDatePicker={showSidebar} />
+
       <div className='page__content'>
-        <div className='wrapper__sidebar'>
-          <Sidebar
-            currentPath={location.pathname}
-            params={params}
-          />
-          <div className='page__content--shortened'>
+        <div className="wrapper__sidebar">
+          {showSidebar && <Sidebar currentPath={pathname} params={params} />}
+          <div
+            className={
+              showSidebar ? 'page__content--shortened' : 'page__content'
+            }
+          >
             <Switch>
               <Route exact path='/reconciliation-reports' render={(props) => <ReconciliationReportList queryParams={filteredQueryParams} {...props} />} />
+              <Route path="/reconciliation-reports/create" component={CreateReconciliationReport} />
               <Route path='/reconciliation-reports/report/:reconciliationReportName' component={ReconciliationReport} />
             </Switch>
-            <section className='page__section'>
-              <Legend />
-            </section>
+            {showSidebar
+              ? (
+                <section className='page__section'>
+                  <Legend />
+                </section>
+              )
+              : null}
           </div>
         </div>
       </div>