+++ app/src/js/components/TextAreaForm/text.js
@@ -18,7 +18,8 @@ class TextForm extends React.Component {
       id,
       error,
       type,
-      className
+      className,
+      ...rest
     } = this.props;
 
     type = type || 'text';
@@ -28,6 +29,7 @@ class TextForm extends React.Component {
         <label htmlFor={id}>{label}</label>
         {error && <span className='form__error'>{error}</span>}
         <input
+          {...rest}
           id={id}
           type={type}
           value={value}