+++ test/components/reconciliation-reports/reconciliation-report.js
@@ -11,12 +11,49 @@ configure({ adapter: new Adapter() });
 
 const reconciliationReports = {
   map: {
-    exampleReport: {
+    exampleInventoryReport: {
       data: {
         reportStartTime: '2018-06-11T18:52:37.710Z',
         reportEndTime: '2018-06-11T18:52:39.893Z',
         status: 'SUCCESS',
         error: null,
+        reportType: 'Inventory',
+        okFileCount: 21,
+        filesInCumulus: {
+          okCount: 129,
+          onlyInS3: [
+            's3://some-bucket/path/to/key-1.hdf',
+            's3://some-bucket/path/to/key-2.hdf'
+          ],
+          onlyInDynamoDb: [
+            {
+              uri: 's3://some-bucket/path/to/key-123.hdf',
+              granuleId: 'g-123'
+            },
+            {
+              uri: 's3://some-bucket/path/to/key-456.hdf',
+              granuleId: 'g-456'
+            }
+          ],
+        },
+        collectionsInCumulusCmr: {
+          okCount: 1
+        },
+        granulesInCumulusCmr: {
+          okCount: 7
+        },
+        filesInCumulusCmr: {
+          okCount: 4
+        }
+      }
+    },
+    exampleGranuleNotFoundReport: {
+      data: {
+        createStartTime: '2018-06-11T18:52:37.710Z',
+        createEndTime: '2018-06-11T18:52:39.893Z',
+        status: 'SUCCESS',
+        error: null,
+        reportType: 'Granule Not Found',
         okFileCount: 21,
         filesInCumulus: {
           okCount: 129,
@@ -57,8 +94,8 @@ const reconciliationReports = {
   }
 };
 
-test('shows an individual report', function (t) {
-  const match = { params: { reconciliationReportName: 'exampleReport' } };
+test('shows an individual inventory report', function (t) {
+  const match = { params: { reconciliationReportName: 'exampleInventoryReport' } };
 
   const dispatch = () => {};
 
@@ -72,10 +109,14 @@ test('shows an individual report', function (t) {
 
   t.is(report.length, 1);
 
-  const ReportHeading = report.find('ReportHeading');
+  const InventoryReport = report.find('InventoryReport');
+  t.is(InventoryReport.length, 1);
+  const inventoryReportWrapper = InventoryReport.dive();
+
+  const ReportHeading = inventoryReportWrapper.find('ReportHeading');
   t.is(ReportHeading.length, 1);
 
-  const TableCards = report.find('TableCards');
+  const TableCards = inventoryReportWrapper.find('TableCards');
   t.is(TableCards.length, 2);
   const TableCardWrapper = TableCards.at(0).dive();
   const Cards = TableCardWrapper.find('Card');
@@ -83,13 +124,12 @@ test('shows an individual report', function (t) {
   // there should be one card for DynamoDB and one card for S3
   t.is(Cards.length, 2);
 
-  const Table = report.find('SortableTable');
+  const Table = inventoryReportWrapper.find('SortableTable');
   t.is(Table.length, 1);
 });
 
-test('correctly renders the heading', function (t) {
-
-  const match = { params: { reconciliationReportName: 'exampleReport' } };
+test('shows an individual Granule Not Found report', function (t) {
+  const match = { params: { reconciliationReportName: 'exampleGranuleNotFoundReport' } };
 
   const dispatch = () => {};
 
@@ -103,7 +143,38 @@ test('correctly renders the heading', function (t) {
 
   t.is(report.length, 1);
 
-  const ReportHeading = report.find('ReportHeading');
+  const GnfReport = report.find('GnfReport');
+  t.is(GnfReport.length, 1);
+  const gnfReportWrapper = GnfReport.dive();
+
+  const ReportHeading = gnfReportWrapper.find('ReportHeading');
+  t.is(ReportHeading.length, 1);
+
+  const Table = gnfReportWrapper.find('SortableTable');
+  t.is(Table.length, 1);
+});
+
+test('correctly renders the heading', function (t) {
+
+  const match = { params: { reconciliationReportName: 'exampleInventoryReport' } };
+
+  const dispatch = () => {};
+
+  const report = shallow(
+    <ReconciliationReport
+      match={match}
+      reconciliationReports={reconciliationReports}
+      dispatch={dispatch}
+    />
+  );
+
+  t.is(report.length, 1);
+
+  const InventoryReport = report.find('InventoryReport');
+  t.is(InventoryReport.length, 1);
+  const inventoryReportWrapper = InventoryReport.dive();
+
+  const ReportHeading = inventoryReportWrapper.find('ReportHeading');
   t.is(ReportHeading.length, 1);
 
   const { downloadOptions, ...headingProps } = ReportHeading.props();
@@ -111,9 +182,10 @@ test('correctly renders the heading', function (t) {
   const expectedHeadingProps = {
     endTime: '2018-06-11T18:52:39.893Z',
     error: null,
-    name: 'exampleReport',
+    name: 'exampleInventoryReport',
     reportState: 'CONFLICT',
     startTime: '2018-06-11T18:52:37.710Z',
+    type: 'Inventory'
   };
 
   t.is(downloadOptions.length, 2);
@@ -141,9 +213,12 @@ test('report with error triggers error message', function (t) {
 
   t.is(report.length, 1);
 
-  const ReportHeading = report.find('ReportHeading');
-  t.is(ReportHeading.length, 1);
+  const InventoryReport = report.find('InventoryReport');
+  t.is(InventoryReport.length, 1);
+  const inventoryReportWrapper = InventoryReport.dive();
 
+  const ReportHeading = inventoryReportWrapper.find('ReportHeading');
+  t.is(ReportHeading.length, 1);
   const reportHeadingWrapper = ReportHeading.dive();
 
   const ErrorReport = reportHeadingWrapper.find('ErrorReport');