+++ CHANGELOG.md
@@ -19,11 +19,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Added TableHeader component to allow user to select page and limit params
   - Removed all Results Per Page dropdowns in favor of the header
 
+- **CUMULUS-1966**
+  - Add form fields for new params to create report form
+  - Upgrades Cypress to 5.2.0
+
 - **CUMULUS-2046**
   - Add dashboard version to footer
 
 - **CUMULUS-2076**
   - Add context references in headings on indivdual pages for screen reader accessibility
+
+- **CUMULUS-2087**
+  - Add Reconciliation Report creation page with report type selection
+
+- **CUMULUS-2089**
+  - Add component for Granule Not Found reports
+
+- **CUMULUS-2114**
+  - Add download button to the Internal report link
 
 ### Changed